requirements.txt
streamlit>=1.38.0
Pillow>=10.4.0
google-generativeai>=0.8.3
gTTS>=2.5.1
requests>=2.32.3
